+++ Assets/CardboardCore/EntityComponents/Entity.cs
@@ -68,7 +68,7 @@ namespace CardboardCore.EntityComponents
             }
         }
 
-        public void Update(double deltaTime)
+        public void Update(float deltaTime)
         {
             for (int i = 0; i < components.Count; i++)
             {
